+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog - golicense
 
+## Unreleased
+
+- update to display the license report by default
+- added the `dump` option (`--dump`) to dump all license files
+
 ## 0.1.0 - 07 February 2022
 
 - First release